refactor: Initial refactor for inv4_git

+++ src/primitives.rs
@@ -9,8 +9,8 @@ pub struct Settings {
 
 #[derive(Debug)]
 pub struct Key {
-    pub ips_id: u64,
-    pub subasset_id: Option<u64>,
+    pub ips_id: u32,
+    pub subasset_id: Option<u32>,
 }
 
 pub type BoxResult<T> = Result<T, Box<dyn Error>>;
